refactor: term sealed class
