+++ exact-numbers/src/main/kotlin/xyz/lbres/exactnumbers/expressions/term/Term.kt
@@ -7,111 +7,43 @@ import xyz.lbres.exactnumbers.common.castToFloat
 import xyz.lbres.exactnumbers.common.castToInt
 import xyz.lbres.exactnumbers.common.castToLong
 import xyz.lbres.exactnumbers.common.castToShort
-import xyz.lbres.exactnumbers.common.createHashCode
 import xyz.lbres.exactnumbers.common.deprecatedV1
-import xyz.lbres.exactnumbers.common.divideByZero
 import xyz.lbres.exactnumbers.common.irrationalPackage
 import xyz.lbres.exactnumbers.exactfraction.ExactFraction
-import xyz.lbres.exactnumbers.ext.divideBy
 import xyz.lbres.exactnumbers.irrationals.common.IrrationalNumber
 import xyz.lbres.exactnumbers.irrationals.log.Log
 import xyz.lbres.exactnumbers.irrationals.pi.Pi
 import xyz.lbres.exactnumbers.irrationals.sqrt.Sqrt
 import xyz.lbres.kotlinutils.collection.ext.toConstMultiSet
 import xyz.lbres.kotlinutils.general.simpleIf
-import xyz.lbres.kotlinutils.set.multiset.const.ConstMultiSet
-import xyz.lbres.kotlinutils.set.multiset.const.emptyConstMultiSet
-import xyz.lbres.kotlinutils.set.multiset.filterConsistent
-import xyz.lbres.kotlinutils.set.multiset.mapToSetConsistent
 import java.math.BigDecimal
 import kotlin.math.abs
 
 /**
  * Representation of the product of several numbers, as a rational coefficient and list of irrational numbers
  */
-class Term private constructor(coefficient: ExactFraction, irrationals: ConstMultiSet<IrrationalNumber<*>>) : Number() {
-    val coefficient: ExactFraction
+sealed class Term : Number() {
+    abstract val coefficient: ExactFraction
+    abstract val irrationals: List<IrrationalNumber<*>>
+    abstract val logs: List<Log>
+    abstract val squareRoots: List<Sqrt>
+    abstract val pis: List<Pi>
+    abstract val piCount: Int
 
-    private val irrationalTypes: MutableMap<String, List<IrrationalNumber<*>>> = mutableMapOf()
+    abstract operator fun unaryMinus(): Term
+    abstract operator fun unaryPlus(): Term
 
-    // TODO finalize property names
-    private val _irrationals: ConstMultiSet<IrrationalNumber<*>>
-    val irrationals: List<IrrationalNumber<*>>
+    abstract operator fun times(other: Term): Term
+    abstract operator fun div(other: Term): Term
 
-    @Suppress("UNCHECKED_CAST")
-    val logs: List<Log>
-        get() = getIrrationalsByType(Log.TYPE) as List<Log>
-
-    @Suppress("UNCHECKED_CAST")
-    val squareRoots: List<Sqrt>
-        get() = getIrrationalsByType(Sqrt.TYPE) as List<Sqrt>
-
-    @Suppress("UNCHECKED_CAST")
-    val pis: List<Pi>
-        get() = getIrrationalsByType(Pi.TYPE) as List<Pi>
-
-    val piCount: Int
-        get() = calculatePiCount()
-
-    // previously computed values for method returns
-    private var simplified: Term? = null
-    private var value: BigDecimal? = null
-    private var string: String? = null
-
-    init {
-        if (coefficient.isZero() || irrationals.any { it.isZero() }) {
-            this.coefficient = ExactFraction.ZERO
-            this.irrationals = emptyList()
-            this._irrationals = emptyConstMultiSet()
-        } else {
-            this.coefficient = coefficient
-            this.irrationals = irrationals.toList()
-            this._irrationals = irrationals
-        }
-    }
-
-    operator fun unaryMinus(): Term = Term(-coefficient, _irrationals)
-    operator fun unaryPlus(): Term = Term(coefficient, _irrationals)
-
-    override fun equals(other: Any?): Boolean {
-        if (other == null || other !is Term) {
-            return false
-        }
-
-        val simplified = getSimplified()
-        val otherSimplified = other.getSimplified()
-
-        return simplified.coefficient == otherSimplified.coefficient && simplified._irrationals == otherSimplified._irrationals
-    }
-
-    operator fun times(other: Term): Term {
-        val newIrrationals = _irrationals + other._irrationals
-        return Term(coefficient * other.coefficient, newIrrationals.toConstMultiSet())
-    }
-
-    operator fun div(other: Term): Term {
-        if (other.isZero()) {
-            throw divideByZero
-        }
-
-        val newIrrationals = _irrationals + other._irrationals.mapToSetConsistent { it.inverse() }
-        return Term(coefficient / other.coefficient, newIrrationals.toConstMultiSet())
-    }
-
-    fun isZero(): Boolean = coefficient.isZero()
+    abstract fun isZero(): Boolean
 
     /**
-     * Simplify all numbers, based on the simplify function for their type
+     * Simplify list of numbers and coefficient
      *
      * @return [Term] simplified version of this term
      */
-    fun getSimplified(): Term {
-        if (simplified == null) {
-            simplified = simplifyTerm(this)
-        }
-
-        return simplified!!
-    }
+    abstract fun getSimplified(): Term
 
     /**
      * Get value of term by multiplying numbers.
@@ -119,19 +51,7 @@ class Term private constructor(coefficient: ExactFraction, irrationals: ConstMul
      *
      * @return [BigDecimal]
      */
-    fun getValue(): BigDecimal {
-        if (value == null) {
-            val simplified = getSimplified()
-
-            val irrationalProduct = simplified.irrationals.fold(BigDecimal.ONE) { acc, number -> acc * number.getValue() }
-            val numeratorProduct = simplified.coefficient.numerator.toBigDecimal() * irrationalProduct
-
-            val result = numeratorProduct.divideBy(simplified.coefficient.denominator.toBigDecimal())
-            value = result
-        }
-
-        return value!!
-    }
+    abstract fun getValue(): BigDecimal
 
     /**
      * Get list of irrational numbers with a given type
@@ -139,51 +59,16 @@ class Term private constructor(coefficient: ExactFraction, irrationals: ConstMul
      * @param type [String]: type to retrieve numbers for
      * @return [List]<IrrationalNumber<*>>: list of irrational numbers, which all have type [type]
      */
-    fun getIrrationalsByType(type: String): List<IrrationalNumber<*>> {
-        if (type !in irrationalTypes) {
-            val result = _irrationals.filterConsistent { it.type == type }
-            irrationalTypes[type] = result
-        }
+    abstract fun getIrrationalsByType(type: String): List<IrrationalNumber<*>>
 
-        return irrationalTypes[type]!!
-    }
+    override fun toByte(): Byte = castToByte(getValue(), this, "Term")
+    override fun toChar(): Char = castToChar(getValue(), this, "Term")
+    override fun toShort(): Short = castToShort(getValue(), this, "Term")
+    override fun toInt(): Int = castToInt(getValue(), this, "Term")
+    override fun toLong(): Long = castToLong(getValue(), this, "Term")
 
-    /**
-     * Calculate number of pis based on list of irrationals
-     *
-     * @return [Int]: number of pis
-     */
-    private fun calculatePiCount(): Int {
-        @Suppress("UNCHECKED_CAST")
-        val pis = getIrrationalsByType(Pi.TYPE) as List<Pi>
-        val positive = pis.count { !it.isInverted }
-        val negative = pis.size - positive
-        return positive - negative
-    }
-
-    override fun toString(): String {
-        if (string == null) {
-            val fractionString = coefficient.toFractionString()
-            val coeffString = simpleIf(fractionString.contains("/"), "[$fractionString]", fractionString)
-            val numString = _irrationals.joinToString("x")
-            val result = simpleIf(numString.isEmpty(), "<$coeffString>", "<${coeffString}x$numString>")
-
-            string = result
-        }
-
-        return string!!
-    }
-
-    override fun toByte(): Byte = castToByte(getValue(), this)
-    override fun toChar(): Char = castToChar(getValue(), this)
-    override fun toShort(): Short = castToShort(getValue(), this)
-    override fun toInt(): Int = castToInt(getValue(), this)
-    override fun toLong(): Long = castToLong(getValue(), this)
-
-    override fun toFloat(): Float = castToFloat(getValue(), this)
-    override fun toDouble(): Double = castToDouble(getValue(), this)
-
-    override fun hashCode(): Int = createHashCode(listOf(coefficient, _irrationals, this::class.toString()))
+    override fun toFloat(): Float = castToFloat(getValue(), this, "Term")
+    override fun toDouble(): Double = castToDouble(getValue(), this, "Term")
 
     @Deprecated("Method $deprecatedV1", ReplaceWith("getIrrationalsByType(Log.TYPE)", "$irrationalPackage.log.Log"), DeprecationLevel.WARNING)
     @JvmName("getLogsDeprecated")
@@ -198,8 +83,8 @@ class Term private constructor(coefficient: ExactFraction, irrationals: ConstMul
     fun getSquareRoots(): List<Sqrt> = squareRoots
 
     companion object {
-        val ZERO = Term(ExactFraction.ZERO, emptyConstMultiSet())
-        val ONE = Term(ExactFraction.ONE, emptyConstMultiSet())
+        val ZERO = fromValues(ExactFraction.ZERO, emptyList())
+        val ONE = fromValues(ExactFraction.ONE, emptyList())
 
         /**
          * Construct a term by providing information about coefficient and irrationals
@@ -209,7 +94,7 @@ class Term private constructor(coefficient: ExactFraction, irrationals: ConstMul
          * @return [Term] with the given values
          */
         fun fromValues(coefficient: ExactFraction, irrationals: List<IrrationalNumber<*>>): Term {
-            return Term(coefficient, irrationals.toConstMultiSet())
+            return TermImpl(coefficient, irrationals.toConstMultiSet())
         }
 
         /**
